+++ webui/react/src/hooks/useModal/UserSettings/useModalChangeName.tsx
@@ -5,7 +5,7 @@ import { StoreAction, useStore, useStoreDispatch } from 'contexts/Store';
 import { patchUser } from 'services/api';
 import handleError from 'utils/error';
 
-import useModal, { ModalHooks } from './useModal';
+import useModal, { ModalHooks } from '../useModal';
 import css from './useModalChangeName.module.scss';
 
 const useModalChangeName = (): ModalHooks => {
