+++ webui/react/src/hooks/useModal/UserSettings/useModalChangePassword.tsx
@@ -5,7 +5,7 @@ import { useStore } from 'contexts/Store';
 import { login, setUserPassword } from 'services/api';
 import handleError from 'utils/error';
 
-import useModal, { ModalHooks } from './useModal';
+import useModal, { ModalHooks } from '../useModal';
 import css from './useModalChangePassword.module.scss';
 
 const useModalChangePassword = (): ModalHooks => {
